Final main commit
With this commit the vast majority of 2-2 is now manually translated. The only thing left is the alternate boss fight with the Twins, but since that's something optional you can do after finishing chapter 2, I've left it for now. The main components of this commit are:

* All of Mother Rabbit's "My Room" dialogue barring the engagement
* Some fix-ups of previous translations
* Some more notes
* The post-Mukuro boss fight dialogue

Also left mostly untranslated is that wet note found in the underground mall. I'm not fluent enough to fill in the missing gaps so I have absolutely no idea what it's supposed to say. There's also a few "Iffy" translations floating around, some because I wasn't 100% on the meaning, but also a lot of them I had trouble phrasing them elegantly into English.